+++ java/code/src/com/redhat/rhn/frontend/action/channel/TargetSystemsConfirmAction.java
@@ -105,7 +105,7 @@ public class TargetSystemsConfirmAction extends RhnAction implements Listable {
      */
     public List getResult(RequestContext context) {
         User user =  context.getLoggedInUser();
-        Long cid = (Long) context.getRequiredParam(RequestContext.CID);
+        Long cid = context.getRequiredParam(RequestContext.CID);
         Channel chan = ChannelManager.lookupByIdAndUser(cid, user);
         return SystemManager.inSet(user, TargetSystemsAction.getSetDecl(chan).getLabel());
     }
